requirements.txt
torch>=1.10
numpy>=1.23
pandas>=1.5
PyYAML>=6.0
tqdm>=4.64
scikit-learn>=1.2
